refactor: QueryService로 변경, 패키지 구조 변경 - #12
